moving reducer in a dedicated file
